requirements.txt
requests
pytest
selenium
pyyaml